[MSPAINT] Do type-cast (BYTE) for lfItalic and lfUnderline
CORE-19094

+++ base/applications/mspaint/textedit.cpp
@@ -340,8 +340,8 @@ void CTextEditWindow::UpdateFont()
     ZeroMemory(&lf, sizeof(lf));
     lf.lfCharSet = DEFAULT_CHARSET; // registrySettings.CharSet; // Ignore
     lf.lfWeight = (registrySettings.Bold ? FW_BOLD : FW_NORMAL);
-    lf.lfItalic = registrySettings.Italic;
-    lf.lfUnderline = registrySettings.Underline;
+    lf.lfItalic = (BYTE)registrySettings.Italic;
+    lf.lfUnderline = (BYTE)registrySettings.Underline;
     lstrcpyn(lf.lfFaceName, registrySettings.strFontName, _countof(lf.lfFaceName));
 
     HDC hdc = GetDC();
